requirements.txt
fastapi==0.95.2
uvicorn==0.24.0
python-multipart==0.0.6
SpeechRecognition==3.10.0
pyaudio==0.2.13
nltk==3.8.1
spacy==3.5.3
openai==1.12.0
python-dotenv==1.0.1
pydantic==1.10.13
numpy==1.24.3
torch==2.1.1
transformers==4.37.2
soundfile==0.12.1
librosa==0.10.1
deep-translator==1.11.4
websockets==12.0 